refactor: split tech tree, technology panel, fleet tile, dialogue view (#3878)
Extract library parts for tech-tree catalog/canvas assembly, technology
panel Slots-tab body, fleet expansion title row chrome, and CtDialogueView
Jenny/collapsible hooks without changing public APIs or player-facing behavior.

Refs #3878
