+++ app/lib/features/game/widgets/panels/fleet_expansion_tile.dart
@@ -10,6 +10,7 @@ import '../units/shared/units_entity_action_row.dart';
 import '../units/shared/units_entity_card.dart';
 
 part 'fleet_expansion_tile_expanded.dart';
+part 'fleet_expansion_tile_title.dart';
 
 /// Naval-units fleet row.
 ///
@@ -75,92 +76,10 @@ class FleetExpansionTile extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.only(left: CtSpacing.m),
       child: UnitsEntityCard(
-        title: _buildTitle(),
-        subtitle: _buildSubtitle(),
+        title: buildFleetTitleRow(),
+        subtitle: buildFleetSubtitle(),
         children: _buildChildren(),
       ),
-    );
-  }
-
-  Widget _buildTitle() {
-    // `chrome: false`: the surrounding bordered gradient card is supplied by
-    // [UnitsEntityCard] so the action row must not paint its own
-    // [UnitsPanelRowChrome] border (issue #3514 AC-6).
-    return UnitsEntityActionRow(
-      chrome: false,
-      dense: true,
-      details: _buildTitleDetails(),
-      actions: _buildTitleActions(),
-    );
-  }
-
-  Widget _buildTitleDetails() {
-    return Row(
-      children: [
-        Checkbox(
-          value: isSelectedForCombine,
-          onChanged: combineSelectionEnabled
-              ? (_) => onCombineSelectionToggle()
-              : null,
-          visualDensity: VisualDensity.compact,
-          materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-        ),
-        const SizedBox(width: 4),
-        Flexible(child: Text(row.label, overflow: TextOverflow.ellipsis)),
-        if (row.isHomeFleet) ...[
-          const SizedBox(width: 6),
-          _HomeFleetChip(label: l10n.naval_units_homeFleetChip),
-        ],
-      ],
-    );
-  }
-
-  List<UnitsEntityAction> _buildTitleActions() {
-    final actions = <UnitsEntityAction>[];
-    if (onMoveFleet != null) {
-      actions.add(
-        UnitsEntityAction(
-          tooltip: l10n.common_move,
-          icon: Icons.route,
-          label: l10n.common_move,
-          onPressed: onMoveFleet,
-          buttonKey: kCtE2EFleetMoveActionKey,
-        ),
-      );
-    }
-    if (isSplitAllowed) {
-      actions.add(
-        UnitsEntityAction(
-          tooltip: l10n.common_split,
-          icon: Icons.call_split,
-          label: l10n.common_split,
-          onPressed: onSplitFleet,
-          buttonKey: kCtE2EFleetSplitActionKey,
-        ),
-      );
-    }
-    if (onTap != null) {
-      actions.add(
-        UnitsEntityAction(
-          tooltip: l10n.naval_units_locateFleet,
-          icon: Icons.my_location,
-          label: l10n.naval_units_locateFleet,
-          onPressed: onTap,
-          iconOnly: true,
-        ),
-      );
-    }
-    return actions;
-  }
-
-  Widget _buildSubtitle() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(row.locationLabel),
-        Text(l10n.naval_units_mission(row.missionLabel)),
-        if (row.draftNavalMoveLine != null) Text(row.draftNavalMoveLine!),
-      ],
     );
   }
