+++ app/lib/features/game/widgets/technology/tech_tree_widget_catalog.dart
@@ -1,0 +1,29 @@
+// Category color/icon maps for [TechTreeWidget].
+// Split from `tech_tree_widget.dart` to keep the host under the repo
+// file-size target (Refs #3878).
+
+part of 'tech_tree_widget.dart';
+
+/// Category color map. SPEC/ui/tech-tree-widget.md: color-coded by category.
+const Map<String, Color> _categoryColors = {
+  'gathering': Color(0xFF2E7D32),
+  'transport': Color(0xFF1565C0),
+  'labour': Color(0xFFF9A825),
+  'civilian': Color(0xFF6A1B9A),
+  'diplomacy': Color(0xFF00838F),
+  'naval': Color(0xFF0D47A1),
+  'military': Color(0xFFC62828),
+  'new-world': Color(0xFF4E342E),
+};
+
+/// Category icon map. SPEC/ui/tech-tree-widget.md: one icon per category.
+const Map<String, String> _categoryIcons = {
+  'gathering': '${kAppIconAssetPrefix}ui_icon_tech_gathering.png',
+  'new-world': '${kAppIconAssetPrefix}ui_icon_tech_new_world.png',
+  'transport': '${kAppIconAssetPrefix}ui_icon_tech_transport.png',
+  'labour': '${kAppIconAssetPrefix}ui_icon_tech_labour.png',
+  'civilian': '${kAppIconAssetPrefix}ui_icon_tech_civilian.png',
+  'diplomacy': '${kAppIconAssetPrefix}ui_icon_tech_diplomacy.png',
+  'naval': '${kAppIconAssetPrefix}ui_icon_tech_naval.png',
+  'military': '${kAppIconAssetPrefix}ui_icon_tech_military.png',
+};
